% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create map of state FARS events in a year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The numerical code corresponding with a U.S. tate}

\item{year}{The four-digit year of data requested}
}
\value{
A map with accidents plotted by latitutde and longitude
}
\description{
This function creates a map of a US state with locations
of fatal accidents in a given year plotted on it. It requires
specification of \code{state.num} and \code{year} variables.
The state number must exist and the year must be valid.
}
