% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in FARS file based on year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years of data to read in}
}
\value{
A tibble of FARS data with number of events in with each year by month
}
\description{
This function searches for and reads in a FARS data file
with the specified \code{years} list. It then converts the
data into a Tibble with the number of incidents in each month
for each specified year. It will notify if the years are invalid.
The list of years must be valid.
}
