% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data observations by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of numerical years}
}
\value{
A single Tibble comprised of 1 or more years of FARS data
}
\description{
This function creates a single Tibble of all FARS data across
multiple years. It only takes a list of \code{years}. It is a
function within the \code{fars_map_state} and \code{fars_summarize_years}
functions. The years must be a valid list.
}
