% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename for a FARS data file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric or integer corresponding with the data's year}
}
\value{
A character string of the filename in the format accident_year.csv.bz2
}
\description{
This function creates a filename in a standardized format for
FARS data. It only takes the argument \code{year}.
}
